homework sprint 6 fix: класс HistoryStorage вынесен в отдельный файл из-за непрохождения автоматических тестов репозитории
